video-processor/requirements.txt
boto3
opencv-python-headless
python-dotenv
httpx